% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{addresses_sample}
\alias{addresses_sample}
\title{Sample data for addresses}
\format{
\subsection{\code{addresses_sample}}{

A data frame with 12335 rows and 6 columns:
\describe{
\item{id_brut_bv}{BV id}
\item{code_commune_ref}{City code}
\item{longitude}{longitude (WGS84)}
\item{latitude}{latitude (WGS84)}
\item{geo_type}{Type of geolocalization (BAN variable)}
\item{geo_score}{Geolocalization accuracy score (BAN variable)}
\item{nb_adresses}{Number of addresses on each line (= for each geolocalization)}
}
}
}
\source{
Insee
}
\usage{
addresses_sample
}
\description{
A subset of addresses data
}
\keyword{datasets}
