% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contours_com_sample}
\alias{contours_com_sample}
\title{Sample data for city contours}
\format{
\subsection{\code{contours_com_sample}}{

A data frame with 3 rows and 2 columns:
\describe{
\item{code_insee}{City code}
\item{geometry}{sf column (WGS84)}
}
}
}
\source{
IGN
}
\usage{
contours_com_sample
}
\description{
A subset of city contours data (IGN - BDTopo)
}
\keyword{datasets}
