% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_voro.R
\name{count_voro}
\alias{count_voro}
\title{count_voro}
\usage{
count_voro(contours, voronoi, marg = 1e-04)
}
\arguments{
\item{contours}{: multipolygons (union of voronoï polygons)}

\item{voronoi}{: polygons (or multipolygon sometimes)}

\item{marg}{: technical parameter}
}
\value{
number of voronoi polygons per contour (using intersections)
}
\description{
Count the number of voronoï polygons used to create voting areas (contours)
}
\examples{
\dontrun{
#contours <- lcontours$`29199` \%>\% st_cast_bis
#voronoi <- lvoronoi$`29199` \%>\% st_cast_bis
#marg <-  0.0001
#count_voro(contours,voronoi)
}
}
