% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_contours.R
\name{create_contours}
\alias{create_contours}
\title{create vote offices contours}
\usage{
create_contours(
  prep_adr,
  cog,
  min_points_com,
  min_address_bv,
  min_address_shoot,
  var_cog1 = "code_commune_ref",
  var_cog2 = "codgeo",
  var_bv1 = "id_brut_bv",
  var_geo_score = "geo_score",
  var_nbaddress = "nb_adresses",
  path_log = NULL
)
}
\arguments{
\item{prep_adr}{Result from prepare_address function :
\itemize{
\item address : table of geolocalized addresses (all cities of the scope)
\item com : city contours (sf, all scope)
\item lcog_ok : scope of cities available
}}

\item{cog}{official targetted city code}

\item{min_points_com}{minimal number of address by city (global parameter)}

\item{min_address_bv}{minimal number of address by vote office (global parameter)}

\item{min_address_shoot}{minimal number of address by polygons in result contours (global parameter)}

\item{var_cog1}{name for city code variable in address table}

\item{var_cog2}{name for city code variable in city-contours table}

\item{var_bv1}{name for bv code variable in address table}

\item{var_geo_score}{name for geo_score variable in address table}

\item{var_nbaddress}{name for number of addresses for each point in address table}

\item{path_log}{path where to write logs (default NULL)}
}
\value{
\itemize{
\item if city in the scope : list of raw contours and simplified contours
\item NULL otherwise (for looping purpose)
}
}
\description{
\itemize{
\item Crée les contours des bureaux de votes pour une ville donnée
\item Create all outputs needed for a targeted city
}
}
\details{
\itemize{
\item Addresses out of their geographic contours are filtered
\item Only works for cities verifying at least "min_points_com" points
\item Create BV contours only if the BV contains at least "min_address_bv" addresses
\item Creates voronoi polygons (intersected with city borders)
\item Links each Voronoi polygon with one single BV (hypothesis in "decouplage_ptsBv" function documentation)
\item Creates BV contours grouping voronoi polygons by BVs
\item Gives a simplified version of those contours trying to remove little cast polygons
with less than "min_address_shoot" addresses (see "shoot_isolated" documentation)
}
}
\examples{
address <- mapvotr::addresses_sample
contours_com <- mapvotr::contours_com_sample
prep_adr <- prepare_address(
  address = address,
  contours_com = contours_com,
  var_cog1 = "code_commune_ref",
  var_bv1 = "id_brut_bv",
  path_log = NULL
)
cog <- "29039"
path_log <- NULL
min_points_com <- 50
min_address_bv <- 15
min_address_shoot <- 5
var_cog1 <- "code_commune_ref"
var_cog2 <- "codgeo"
var_bv1 <- "id_brut_bv"
var_geo_score <- "geo_score"
var_nbaddress <- "nb_adresses"
create_contours(prep_adr, cog, min_points_com, min_address_bv, min_address_shoot)
}
