% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsg_from_cog.R
\name{epsg_from_cog}
\alias{epsg_from_cog}
\title{Guess EPSG code}
\usage{
epsg_from_cog(cog)
}
\arguments{
\item{cog}{Municipality official code}
}
\value{
epsg code
}
\description{
Guess epsg code using city code (choosing between metropolitan France and overseas territories)
}
\examples{
cog = 29019 
epsg_from_cog(cog)
}
