% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_inContour.R
\name{f_inContour}
\alias{f_inContour}
\title{f_inContour}
\usage{
f_inContour(sfpoints, sfcontour, verbose = FALSE)
}
\arguments{
\item{sfpoints}{points (sf)}

\item{sfcontour}{polygon (sf)}

\item{verbose}{if TRUE : prints information}
}
\value{
sfpoints without points out of sfContour
}
\description{
Exclude points out of a border
}
\examples{
sfpoints = addresses_sample[addresses_sample$code_commune_ref == "29039",]
sfpoints <- sfpoints \%>\%
sf::st_as_sf(
  coords = c("longitude", "latitude"), 
  crs = 4326,
  na.fail = FALSE
)
sfcontour = contours_com_sample[contours_com_sample$codgeo == "29039",]
verbose=TRUE
mapvotr:::f_inContour(sfpoints,sfcontour,verbose)
}
