% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_contours.R
\name{map_contours}
\alias{map_contours}
\title{Map your contours}
\usage{
map_contours(
  cog,
  sfelecteurs,
  contours_bv,
  var_code_bv = "id_brut_bv_reu",
  var_score = "geo_score",
  var_cog = "code_commune_ref"
)
}
\arguments{
\item{cog}{Municipality official code}

\item{sfelecteurs}{addresses of the targeted city}

\item{contours_bv}{contours of the targeted city}

\item{var_code_bv}{name of column : id BV}

\item{var_score}{name of column : quality of geocoding}

\item{var_cog}{name of column : official city code (COG)}
}
\value{
dynamic map (leaflet style)
}
\description{
\itemize{
\item Crée des cartes dynamiques des contours
\item Create a dynamic map of contours
}
}
\examples{
\dontrun{
cog <- city_code
sfelecteurs=lres$addresses
contours_bv=lcontours$contours_simplified
var_cog = "code_commune_ref"
var_score="geo_score"
var_code_bv="id_brut_bv_reu"
palettecouleurs="Paired"
map_contours(cog,sfelecteurs,contours_bv)
}
}
