% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_address.R
\name{prepare_address}
\alias{prepare_address}
\title{prepare_address}
\usage{
prepare_address(
  address,
  contours_com,
  var_cog1 = "code_commune_ref",
  var_cog2 = "codgeo",
  var_bv1 = "id_brut_bv",
  path_log = NULL
)
}
\arguments{
\item{address}{address}

\item{contours_com}{city contours}

\item{var_cog1}{name for city code variable in address table}

\item{var_cog2}{name for city code variable in city-contours table}

\item{var_bv1}{name for bv code variable in address table}

\item{path_log}{path to write logs (optional)}
}
\value{
List of :
\itemize{
\item table of address
\item table of cities (geographic contours)
\item List of city-codes (with at least 2 bv)
}
}
\description{
: prepare :
\itemize{
\item a table of address (only those "properly" geolocalized )
\item list of cities with at least 2 BVs
\item List of city contours
}
}
\details{
:
\itemize{
\item only BAN quality "housenumber","interpolation","locality"
\item only address for cities with at leat 2 BVs
}
}
\examples{
addresses_sample <- mapvotr::addresses_sample
contours_com_sample <- mapvotr::contours_com_sample
prepare_address <- prepare_address(address = addresses_sample,
contours_com = contours_com_sample,
var_cog1 = "code_commune_ref",
var_bv1 = "id_brut_bv",
path_log=NULL)
}
