% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_input_voronoi}
\alias{sf_input_voronoi}
\title{Example data for sf_input_voronoi}
\format{
\subsection{\code{sf_input_voronoi}}{

A sf table with 269 rows and 5 columns:
\describe{
\item{id_brut_bv}{BV id}
\item{code_commune_ref}{City code}
\item{geo_type}{Type of geolocalization (BAN variable)}
\item{geo_score}{Geolocalization accuracy score (BAN variable)}
\item{nb_adresses}{Number of addresses on each line (= for each geolocalization)}
\item{geometry}{sf column (WGS84)}
}
}
}
\source{
IGN
}
\usage{
sf_input_voronoi
}
\description{
A subset of real input
}
\keyword{datasets}
