% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoot_isolated.R
\name{shoot_isolated}
\alias{shoot_isolated}
\title{Simplify voting contours}
\usage{
shoot_isolated(contours, voronoi, min_address_shoot, var_bv = "id_brut_bv")
}
\arguments{
\item{contours}{contours}

\item{voronoi}{voronoi polygons}

\item{min_address_shoot}{minimal number of address by polygons in result contours (global parameter)}

\item{var_bv}{name for bv code variable in address table}
}
\value{
simplified contours
}
\description{
Simplify contours removing isolated polygons
}
\details{
\itemize{
\item Counts the number of voronoï polygons used to create each BV contour
\item Targets cast BV Contours (cBVc) built with less than min_address_shoot voronois/points
\item if cBVc is not isolated (island for example) and at least one neighbor is big enough:
\itemize{
\item replace old BV code by the biggest neighbor 's BV code.
}
\item Reshape simplified BV contours
}
}
