% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_for_contours.R
\name{valid_for_contours}
\alias{valid_for_contours}
\title{Check if the city can be divided into voting contours}
\usage{
valid_for_contours(
  tab_address,
  min_points_com,
  min_address_bv,
  var_bv1,
  var_nbaddress
)
}
\arguments{
\item{tab_address}{addresses of a single city}

\item{min_points_com}{minimal number of points by city (global parameter)}

\item{min_address_bv}{minimal number of address by vote office (global parameter)}

\item{var_bv1}{name for bv code variable in address table}

\item{var_nbaddress}{name for number of addresses for each point in address table}
}
\value{
\itemize{
\item TRUE if the city is in the scope,
\item otherwise FALSE or a list of problematic BVs
}
}
\description{
Check if a city can be in the scope (ok to be divided into several geographical contours)
}
\details{
The function uses the addresses of the city targeted
}
\examples{
tab_address <- mapvotr::addresses_sample
min_points_com <- 50
min_address_bv <- 15
var_bv1 = "id_brut_bv_reu"
var_nbaddress = "nb_adresses"
mapvotr:::valid_for_contours(tab_address, min_points_com, min_address_bv,var_bv1,var_nbaddress)

# With some problematic BVs
lines_pb <- tab_address[tab_address$id_brut_bv_reu == "29039_1",][1:5,]
tab_address_pb <- tab_address[tab_address$id_brut_bv_reu != "29039_1",] \%>\% rbind(lines_pb)
mapvotr:::valid_for_contours(tab_address_pb, min_points_com, min_address_bv,var_bv1,var_nbaddress)
}
