% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi_com.R
\name{voronoi_com}
\alias{voronoi_com}
\title{Make Voronoï polygons}
\usage{
voronoi_com(sfelecteurs, com, var_bv)
}
\arguments{
\item{sfelecteurs}{sf points}

\item{com}{city contours (sf multipolygons)}

\item{var_bv}{column name for BV id column}
}
\value{
voronoi multipolygons
}
\description{
Partition a city into voronoi polygons :
\itemize{
\item Use sf::st_voronoi with \code{envelope} parameter (= city contour)
\item Cast the result and intersects with the city contour
}
}
\details{
Create voronoi polygons inside city contours.
\itemize{
\item 1 point = several addresses = several BVs
}
}
\examples{
sfelecteurs <- mapvotr::sf_input_voronoi
com <- mapvotr::contours_com_sample \%>\%
  dplyr::filter(codgeo == "29039") \%>\%
  sf::st_transform(2154)
var_bv <- "id_brut_bv"
mapvotr:::voronoi_com(sfelecteurs, com, var_bv)
}
