% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fget_multiBV.R
\name{fget_multiBV}
\alias{fget_multiBV}
\title{fget_multiBV}
\usage{
fget_multiBV(tab_addresses, var_cog, var_bv)
}
\arguments{
\item{tab_addresses}{addresses}

\item{var_cog}{column name for cog}

\item{var_bv}{column name for BD id}
}
\value{
vector of city codes (for cities with at least 2 vote offices)
}
\description{
Get COG for cities with at leat 2 vote offices
}
\examples{
var_cog = "code_commune_ref"
var_bv = "id_brut_bv"
mapvotr:::fget_multiBV(mapvotr::addresses_sample,"code_commune_ref","id_brut_bv")
}
