% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_address.R
\name{prepare_address}
\alias{prepare_address}
\title{Pre-process addresses}
\usage{
prepare_address(
  address,
  contours_com,
  var_cog1 = "code_commune_ref",
  var_cog2 = "codgeo",
  var_bv1 = "id_brut_bv",
  path_log = NULL
)
}
\arguments{
\item{address}{addresses from REU data (Insee)}

\item{contours_com}{city contours (ex : BDTopo IGN)}

\item{var_cog1}{name for city code variable in address table}

\item{var_cog2}{name for city code variable in city-contours table}

\item{var_bv1}{name for bv code variable in address table}

\item{path_log}{path to write logs (optional)}
}
\value{
List of :
\itemize{
\item table of address
\item table of cities (geographic contours)
\item List of city-codes (with at least 2 bv)
}
}
\description{
\itemize{
\item Prétraitement des données à partir des adresses du REU et des contours de communes
\item Prepares your data from REU addresses and city contours
}
}
\details{
prepares :
\itemize{
\item a table of address
\itemize{
\item scope : only those with a minimal geolocalization quality
\item BAN quality "housenumber", "interpolation" or "locality"
}
\item list of cities with at least 2 BVs (among filtered addresses)
\item List of city geometric contours
}

Operations :
\itemize{
\item Removes arrondissement codes for Paris, Lyon and Marseille
\item Filters addresses (only "housenumber", "interpolation" or "locality" quality)
\item Filters cities with at least 2 vote offices
\item Transform address as geometric points (WGS84 projection)
\item Filters geometric city contours (only cities with at least 2 vote offices)
\item Logs missing city contours
}
}
\examples{
addresses_sample <- mapvotr::addresses_sample
contours_com_sample <- mapvotr::contours_com_sample
prepare_address <- prepare_address(
  address = addresses_sample,
  contours_com = contours_com_sample,
  var_cog1 = "code_commune_ref",
  var_bv1 = "id_brut_bv",
  path_log = NULL
)
}
