% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_arrond.R
\name{rm_arrond}
\alias{rm_arrond}
\title{rm_arrond}
\usage{
rm_arrond(tab, var_cog)
}
\arguments{
\item{tab}{table with a cog column}

\item{var_cog}{name of the cog column}
}
\value{
tab with modified cog column (Paris, Lyon and Marseilles codes without arrondissement)
}
\description{
replace "arrondissements" cog by city cog for Paris, Lyon and Marseille
}
\examples{
tab <- data.frame(id=1:5,code_commune_ref=c("75118","75000","35000","69382",NA))
var_cog <- "code_commune_ref"
mapvotr:::rm_arrond(tab,var_cog)
}
